split out the grid implementations
